requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Login==0.6.2
WTForms==3.0.1
Werkzeug==2.0.1
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
python-dotenv==0.19.0
email-validator==2.0.0
Flask-Mail==0.9.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
Flask-Babel==2.0.0
gunicorn==20.1.0
pytest==7.4.0
coverage==7.2.7